\name{chronoplotR-package}
\alias{chronoplotR-package}
\alias{chronoplotR}
\docType{package}
\title{
\packageTitle{chronoplotR}
}
\description{
\packageDescription{chronoplotR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{chronoplotR}
\packageIndices{chronoplotR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{chronoplotR}

Maintainer: \packageMaintainer{chronoplotR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
