% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleplot.R
\encoding{UTF-8}
\name{doubleplot}
\alias{doubleplot}
\title{Actogram Plot
Representation of animal activities rhythms plot}
\arguments{
\item{timestamp}{a vector of data and time in time stamp format ("yyyy-mm-hh:mm:ss").}

\item{rythm}{a numeric vector related to activities variable for each timestamp.}

\item{n_lines}{number of lines}

\item{dark_beg}{dark begins}

\item{dark_end}{dark ends}

\item{cold_beg}{cold begins}

\item{cold_end}{cold ends}
}
\value{
an doble actogram plot and ritm table
}
\description{
Actogram Plot
Representation of animal activities rhythms plot
}
\section{Details}{
  O argumento n_lines pode ser utilizado para selecionar um conjunto dos dados desejados, tendo-se em mente que como trata-se de dados temporais em sequência, n_lines se refere à quantidade de observações contidas no intervalo 1:n_lines, ou seja, sempre a partir da primeira observação fornecida.
}

\section{References}{
   MARQUES, M. D.; GOLOMBEK, D.; MORENO, C. Adaptação Temporal. In: MARQUES, N. &  MENNA-BARRETO, L. Cronobiologia: princípios e aplicações. São Paulo: Editora Universidade de São Paulo, 1999. p.45-62.
}

\examples{


\dontrun{
day2 <- sort(as.POSIXct("1995-03-31 00:00:00") + runif(n=55, min=0, max=(3600*24)))
day3 <- sort(as.POSIXct("1995-04-01 00:00:00") + runif(n=45, min=0, max=(3600*24)))
day4 <- sort(as.POSIXct("1995-04-02 00:00:00") + runif(n=50, min=0, max=(3600*24)))
day5 <- sort(as.POSIXct("1995-04-03 00:00:00") + runif(n=60, min=0, max=(3600*24)))
day6 <- sort(as.POSIXct("1995-04-04 00:00:00") + runif(n=40, min=0, max=(3600*24)))
day7 <- sort(as.POSIXct("1995-04-05 00:00:00") + runif(n=53, min=0, max=(3600*24)))
day8 <- sort(as.POSIXct("1995-04-06 00:00:00") + runif(n=47, min=0, max=(3600*24)))
day9 <- sort(as.POSIXct("1995-04-07 00:00:00") + runif(n=55, min=0, max=(3600*24)))
day10 <- sort(as.POSIXct("1995-04-08 00:00:00") + runif(n=45, min=0, max=(3600*24)))
timestamp <- c(day1, day2, day3, day4, day5, day6, day7, day8, day9, day10)
rythm <- abs(rnorm(n=500, mean=50, sd=20))
dark_beg <- c(rep("00:00:00",150),rep("06:00:00",110),rep("18:00:00",93),rep("00:00:00",147))
dark_end <- c(rep("00:00:00",150),rep("18:00:00",110),rep("06:00:00",93),rep("23:59:59",147))
cold_beg <- c(rep("00:00:00",150),rep("06:00:00",110),rep("18:00:00",93),rep("04:00:00",147))
cold_end <- c(rep("00:00:00",150),rep("18:00:00",110),rep("06:00:00",93),rep("20:00:00",147))
doubleplot(timestamp=timestamp,rythm=rythm,n_lines=500,dark_beg=dark_beg,dark_end=dark_end,cold_beg=cold_beg,cold_end=cold_end)
doubleplot(timestamp=timestamp,rythm=rythm,n_lines=353,dark_beg=dark_beg,dark_end=dark_end)
doubleplot(timestamp=timestamp,rythm=rythm,n_lines=500,dark_beg=dark_beg,dark_end=dark_end,cold_beg=cold_beg,cold_end=cold_end, bar_width=600, graph_font="AvantGarde", hot_color="lightcoral", x_title_size = 5.5, y_lables_size=2.8, first_day=15, plot_legend=F)
}


}
